% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalue.R
\name{cvalue_eb2s}
\alias{cvalue_eb2s}
\title{Bootstrap Critical Value}
\usage{
cvalue_eb2s(
  x_data,
  alpha,
  beta = alpha/50,
  bootstrap_replications = NULL,
  rng_seed = NULL,
  bootstrap_indices = NULL
)
}
\arguments{
\item{x_data}{an n x k matrix of evaluated moment functions.}

\item{alpha}{the confidence level required for the first stage test.}

\item{beta}{optional confidence level required for the second stage test.}

\item{bootstrap_replications}{the number of bootstrap replications.
Required if bootstrap_indices is not specified.}

\item{rng_seed}{the seed for replication purposes. If not specified, the
seed is not set.}

\item{bootstrap_indices}{an integer vector of indices to use for the
bootstrap. If this is specified, bootstrap_replications and rng_seed will
be ignored. If this is not specified, bootstrap_replications is required.}
}
\value{
a float for the critical value.
}
\description{
Bootstrap Critical Value
}
