% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalue.R
\name{cvalue_sn2s}
\alias{cvalue_sn2s}
\title{Stage 2 Critical Value}
\usage{
cvalue_sn2s(x_data, alpha, beta = alpha/50)
}
\arguments{
\item{x_data}{an n x k matrix of evaluated moment functions.}

\item{alpha}{the confidence level required for the first stage test.}

\item{beta}{optional confidence level required for the second stage test.}
}
\value{
a float for the critical value.
}
\description{
Stage 2 Critical Value
}
