% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{m_hat}
\alias{m_hat}
\title{Standardized Sample Mean of Moment Functions}
\usage{
m_hat(x_data, MARGIN = 2)
}
\arguments{
\item{x_data}{an n x k matrix of evaluated moment functions.}

\item{MARGIN}{a vector giving the subscripts which the function will be
applied over. E.g., for a matrix 1 indicates rows, 2 indicates columns.}
}
\value{
a vector of the standardized sample mean of the moment functions.
}
\description{
Standardized Sample Mean of Moment Functions
}
